+++ src/views/uBikeTable/components/search.vue
@@ -1,0 +1,8 @@
+<script setup>
+    const props = defineProps(['searchText']);
+    const emits = defineEmits(['update:searchText']);
+</script>
+
+<template>
+    <input type="text" class="border" :value="searchText"  @input="emits('update:searchText',$event.target.value)" />
+</template>
